pyproject.toml
[tool.poetry]
name = "delivery_service"
version = "0.1.0"
description = "Микросервис международной доставки"
authors = ["twentytwo-sixteen <gadiev.mingiyan@mail.ru>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
asyncpg = "^0.29.0"
pydantic = "^2.7.1"
aioredis = "^2.0.1"
motor = "^3.4.0"
aiohttp = "^3.9.5"
python-dotenv = "^1.0.1"
pika = "^1.3.2"
aio-pika = "^9.3.0"
celery = "^5.3.6"
loguru = "^0.7.3"
pydantic-settings = "^2.9.1"
alembic = "^1.16.1"
fastapi-cache2 = "^0.2.2"
redis = "^6.2.0"
prometheus-client = "^0.22.1"
tenacity = "^9.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
ruff = "^0.4.4"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
worker = "src.app.rabbitmq.consumer:main"

